Create prosperity
Charmaine is prosperous  in all things she sets out to achieve mentally and physically exerted efforts as well. 
Charmain makes achievements and happiness all the time with what she does and how she does it as she is pleasing to others and they do not seek or think to hurt her or force her into behaving in a way she doesn’t approve of regardless of the past or the past being irrelevant as some know the truth especially the ones that need to know and can keep a secret as only secret keepers does do the dealings of royalty’s lives and Charmain is royalty. It is known she was vocational kidnapped in some places, all the right places and she is being freed easily and happily from this position. 
Charmain and her brother are freed at the same time to make life easy they get to see each other when they are freed and before it if they want it which is likely as they are Idnetical twins it’s a secret that is kept and no one causes harassment or problems because of their relationship and no one attempts to he that for each of them. No one can be their replacement and nothing can satisfy Them the way each other would, could or should and that is a fact they both know and accept and stomach easily. 
Getting Charmain’s identity back is easy as it looks like it has been  stolen, although it is truly impossible to steal an industry some knwo the truth that are qualified to know the truth and are trustworthy in her family; the actions of these individuals do not hinder charmain or cause her stress as they keep their emotions out of their responses and maintain achieving the twins meeting for multiple reasons all describable in sone way and possibly some extras that are not describable and will stay that way until we know how to do that which we will well and truly have met before that time much much before that instance happening. 
Charmain is going into her body and will be clean and healthy and fully functional and won’t be causing any problems for any one in her body so no one has any drama or problem with it their minds are all made up and they do not want to cause harm to Charmain the royal one.